refactor: Make domain package for domain classes
