+++ src/main/java/nextstep/ladder/domain/Ladder.java
@@ -1,9 +1,9 @@
-package nextstep.ladder;
+package nextstep.ladder.domain;
 
 import java.util.List;
 import java.util.stream.Collectors;
 import java.util.stream.IntStream;
-import nextstep.ladder.model.Position;
+import nextstep.ladder.domain.model.Position;
 
 public class Ladder {
