+++ src/main/java/nextstep/ladder/domain/LadderLine.java
@@ -1,12 +1,9 @@
-package nextstep.ladder;
+package nextstep.ladder.domain;
 
 import java.util.HashMap;
-import java.util.List;
 import java.util.Map;
-import java.util.Random;
-import java.util.stream.Collectors;
 import java.util.stream.IntStream;
-import nextstep.ladder.model.Position;
+import nextstep.ladder.domain.model.Position;
 
 public class LadderLine {
