+++ src/main/java/nextstep/ladder/domain/PrizeSheet.java
@@ -1,12 +1,12 @@
-package nextstep.ladder;
+package nextstep.ladder.domain;
 
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.stream.Collectors;
 import java.util.stream.IntStream;
-import nextstep.ladder.model.Position;
-import nextstep.ladder.model.Prize;
+import nextstep.ladder.domain.model.Position;
+import nextstep.ladder.domain.model.Prize;
 
 public class PrizeSheet {
   private Map<Position, Prize> prizeSheet = new HashMap<>();
