+++ src/main/java/nextstep/ladder/domain/model/Name.java
@@ -1,4 +1,4 @@
-package nextstep.ladder.model;
+package nextstep.ladder.domain.model;
 
 import java.util.Objects;
