+++ src/test/java/nextstep/ladder/domain/LadderTest.java
@@ -1,7 +1,9 @@
-package nextstep.ladder;
+package nextstep.ladder.domain;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
+import nextstep.ladder.domain.Ladder;
+import nextstep.ladder.domain.Player;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
