+++ src/test/java/nextstep/ladder/domain/PrizeSheetTest.java
@@ -1,10 +1,12 @@
-package nextstep.ladder;
+package nextstep.ladder.domain;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
 import java.util.Arrays;
 import java.util.List;
 import java.util.stream.IntStream;
+import nextstep.ladder.domain.Player;
+import nextstep.ladder.domain.PrizeSheet;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
