+++ src/test/java/nextstep/ladder/domain/model/PositionTest.java
@@ -1,9 +1,8 @@
-package nextstep.ladder.model;
+package nextstep.ladder.domain.model;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
-import nextstep.ladder.model.Position;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
